% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipquery.R
\name{query_ips}
\alias{query_ips}
\title{Get information about IP Addresses}
\usage{
query_ips(ips)
}
\arguments{
\item{ips}{a character vector of IP addresses. Must not contain NA values.}
}
\value{
A data.frame with columns
\itemize{
\item \code{ip}: character
\item \code{isp}: character
\item \code{asn}: character
\item \code{org}: character
\item \code{country}: character
\item \code{country_code}: character
\item \code{city}: character
\item \code{state}: character
\item \code{zipcode}: character
\item \code{latitude}: numeric
\item \code{longitude}: numeric
\item \code{timezone}: character
\item \code{localtime}: character
\item \code{is_mobile}: logical
\item \code{is_vpn}: logical
\item \code{is_tor}: logical
\item \code{is_proxy}: logical
\item \code{is_datacenter}: logical
\item \code{risk_score}: integer
}
}
\description{
Return detailed information from a vector of IP addresses.
}
\examples{

# Only run this example when internet is present
if (rlang::is_installed("curl")) {
  if (curl::has_internet()) {
    query_ips("1.1.1.1")
  }
}
}
\references{
See \href{https://ipquery.gitbook.io/ipquery-docs}{API Reference}
}
