% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipquery.R
\name{query_own_ip}
\alias{query_own_ip}
\title{Query host IP Address}
\usage{
query_own_ip()
}
\value{
A character string
}
\description{
Returns information about the IP address of the host machine.
}
\examples{
if (rlang::is_installed("curl")) {
  if (curl::has_internet()) {
    query_own_ip()
  }
}
}
